Update vue component
