--- packages/ag-charts-vue3/tsconfig-lib.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es6",
-        "module": "esnext",
-        "moduleResolution": "node",
-        "strict": false,
-        "experimentalDecorators": true,
-        "declaration": true,
-        "rootDir": "src",
-        "outDir": "lib",
-        "skipLibCheck": true
-    },
-    "include": ["src/**/*.ts"],
-    "exclude": ["**/node_modules/*", "src/main.ts"]
-}
