+++ packages/ag-charts-vue3/tsconfig.types.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.lib.json",
+    "compilerOptions": {
+        "declaration": true,
+        "emitDeclarationOnly": true,
+        "types": ["node"]
+    },
+    "include": ["src/*.ts"]
+}
