+++ packages/ag-charts-website/public/example-runner/charts-vue3-boilerplate/systemjs.config.js
@@ -30,8 +30,6 @@
             // vuejs
             vue: 'npm:vue@3.2.29/dist/vue.esm-browser.js',
             '@vue/reactivity': 'npm:@vue/reactivity@3.0.0/dist/reactivity.esm-browser.prod.js',
-            // vue class component
-            'vue-class-component': 'npm:vue-class-component@^8.0.0-beta.3/dist/vue-class-component.cjs.js',
 
             app: appLocation + 'app',
             // systemJsMap comes from index.html
@@ -42,17 +40,11 @@
             vue: {
                 defaultExtension: 'js',
             },
-            'vue-class-component': {
-                defaultExtension: 'js',
-            },
-            'vue-property-decorator': {
-                defaultExtension: 'js',
-            },
             app: {
                 defaultExtension: 'js',
             },
             'ag-charts-vue3': {
-                main: './lib/AgChartsVue.js',
+                main: './dist/package/index.cjs.js',
                 defaultExtension: 'js',
             },
         },
